refactor: 유저의 매수/매도 가능 fetch 함수 분리
